refactor(kill_tor_process) - Alter killing process on mac osx

+++ extractor.py
@@ -30,7 +30,7 @@ logger.set_verbose()
 
 def signalhandler(singnum, frame):
     """Handle a keyboard interrupt"""
-    log('\n[*] Received user keyboard interrupt.')
+    logger.log('\n[*] Received user keyboard interrupt.')
     print('[*] Exiting.')
     sys.exit()
 
@@ -38,7 +38,7 @@ class ProcessNotKilledException(BaseException):
     def __init__(self, message):
         self.message = message
     
-    def __repr__():
+    def __repr__(self):
         return str(self.message)
 
 class Loader(object):
@@ -95,7 +95,7 @@ class Service(object):
         """Try to install service based on user platform"""
         pass
 
-    def execute_process(self, command, shell=True, root=False, loader=None):
+    def execute_process(self, command, shell=True, root=False, loader=None, timeout=5):
         """Handle a process execution returning a tuple with stdout and stderr"""
         if root:
             command[0] = 'sudo ' + command[0]
@@ -107,7 +107,7 @@ class Service(object):
             if root and self._require_root:
                 stdout, stderr = process.communicate()
             else:    
-                stdout, stderr = process.communicate(timeout=5)
+                stdout, stderr = process.communicate(timeout=timeout)
             if loader:
                     loader.start()
         except subprocess.TimeoutExpired:
@@ -174,11 +174,11 @@ class YoutubeDl(Service):
             logger.log('\n[-] Error: '+ stderr, RED)
             logger.log('[*] Last command failed.')
             if platform == 'darwin':
-                command = ['brew', 'install', 'youtube-dl']
-                stderr = self.execute_process(command, root=True)[1]
+                command = ['brew install youtube-dl']
+                stderr = self.execute_process(command, timeout=None)[1]
             elif platform == 'linux':
-                command = ['apt-get', 'install', '-y', 'youtube-dl']
-                stderr = self.execute_process(command, root=True)[1]
+                command = ['apt-get install -y youtube-dl']
+                stderr = self.execute_process(command, timeout=None, root=True)[1]
         loader.stop()
         if stderr:
             logger.log('\n[*] Error installing: '+ stderr, RED)
@@ -197,15 +197,16 @@ class Tor(Service):
        
         logger.log('[*] Starting tor service.')
         if not self._kill_process():
+            self.stop()
             raise ProcessNotKilledException('Could not kill a already running tor process.')
 
         command = ['tor']
         if not self.installed and not self.tried_to_install:
             tries = 0
             while tries < 2:
-                stderr = self.execute_process(command)
+                stderr = self.execute_process(command)[1]
                 tries += 1
-                if stderr:
+                if stderr and stderr != 'Process timed out.':
                     if platform == 'linux':
                         logger.log('[*] Trying to start with systemctl interface.', YELLOW)
                         stderr = self.execute_process(['systemctl start tor.service'], root=True)[1]
@@ -252,8 +253,10 @@ class Tor(Service):
         """Stop running process"""
         logger.log('[*] Stopping tor service.')
         self.event.clear()
-        self._kill_process()
         time.sleep(1)
+
+        # kill any remaining process
+        self._kill_process()
 
     def _start_in_background(self):
         """Start service in background with a Thread"""
@@ -265,13 +268,18 @@ class Tor(Service):
         def run():
             logger.log('[*] Started tor process in background.')
             process = subprocess.Popen(['tor'], shell=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-            self.started = True
             while self.event.is_set():
                 process.communicate()
                 time.sleep(1)
             self.started = False
-            process.kill()
 
+            # receive the process return code
+            if process.poll() and process.poll() < 0:
+                if platform == 'windows':
+                    process.terminate()
+                else:
+                    process.kill()
+        self.started = True
         self.event.set()
         threading.Thread(target=run).start()
     
@@ -281,10 +289,10 @@ class Tor(Service):
         loader.start()
         if platform == 'darwin':
             command = ['brew install tor']
-            stderr = self.execute_process(command, root=True)[1]
+            stderr = self.execute_process(command, timeout=None)[1]
         elif platform == 'linux':
             command = ['apt-get install -y tor']
-            stdout, stderr = self.execute_process(command, root=True)   
+            stdout, stderr = self.execute_process(command, timeout=None, root=True)   
         else:
             loader.stop()
             print('\n[-] You are using Windows. No available installers.')
@@ -315,34 +323,30 @@ class Tor(Service):
                         return True
                 return False
 
-        if platform == 'linux':
-            command = 'sudo systemctl stop tor'
+        elif platform == 'darwin':
+            command = ['lsof -i tcp:9050']
+            stdout, stderr = self.execute_process(command)
+            if not stdout:
+                return True
+            else:
+                stdout = ' '.join([e for e in stdout.split(' ') if e])
+                pid = re.findall(r"[a-zA-Z]\s(\d+)", stdout)
+                if len(pid) > 0:
+                    stderr = self.execute_process(['kill '+ pid[0]], root=True)[1]
+                    if not stderr:
+                        return True
+                return False
+        else:
+            command = ['sudo systemctl stop tor']
             self.execute_process(command)[1]
-
-        command = ['netstat -nlp | grep 9050']
-        stdout, stderr = self.execute_process(command, root=True)
-        if not stdout:
-            return True
-        else:
-            pid = re.findall(r"(\d+)\/[a-zA-Z]+", stdout)
-            if len(pid) > 0:
-                stderr = self.execute_process(['kill '+ pid[0]], root=True)[1]
-                if not stderr:
-                    return True
-            return False
-
-if __name__ == '__main__':
-    # ydl = YoutubeDl()
-    # if ydl.start():
-    #     print('youtube-dl is ready')
-    # else:
-    #     print('youtube-dl is messed up')
-    tor = Tor()
-    if tor.start():
-        print('tor is ready')
-        if tor.restart():
-            print('tor restarted with success')
-        else:
-            print('tor restarted with error')
-    else:
-        print('tor failed on start')
+            command = ['netstat -nlp | grep 9050']
+            stdout, stderr = self.execute_process(command, root=True)
+            if not stdout:
+                return True
+            else:
+                pid = re.findall(r"(\d+)\/[a-zA-Z]+", stdout)
+                if len(pid) > 0:
+                    stderr = self.execute_process(['kill '+ pid[0]], root=True)[1]
+                    if not stderr:
+                        return True
+                return False
